Wrap MeetingRoom component with Suspense to fix useSearchParams CSR bailout error and ensure proper rendering in /room page with Next.js App Router.
